+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 
 ## 2026-04-24
+- `18:17` -- **Updated** `apps/desktopProbe/src/pages/components/quietHoursSettings.tsx`
 - `18:14` -- **Updated** `apps/desktopProbe/src/pages/settings.tsx`
 - `18:14` -- **Updated** `apps/desktopProbe/src/pages/settings.tsx`
 - `18:13` -- **Added** `apps/desktopProbe/src/pages/components/quietHoursSettings.tsx`